refactor(apply): move document model to internal/asset
Extract assetDocument and its reading/parsing helpers (readMultiDocumentYAML,
parseMultiDocumentYAML, readDirectory, discoverFiles, parseDocumentHeader,
formatNameAndId, pluralize) out of internal/apply into internal/asset as
exported Document/ReadMultiDocumentYAML/etc.

apply is the only current caller, but the upcoming diff command needs the
same document discovery and parsing without apply's write-path baggage, and
project convention already places logic shared across apply and per-asset
commands in internal/asset.
