+++ internal/apply/apply.go
@@ -1,13 +1,9 @@
 package apply
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"os"
-	"path/filepath"
-	"sort"
 	"strings"
 
 	dash0api "github.com/dash0hq/dash0-api-client-go"
@@ -18,7 +14,6 @@ import (
 	"github.com/dash0hq/dash0-cli/internal/confirmation"
 	"github.com/dash0hq/dash0-cli/internal/experimental"
 	"github.com/spf13/cobra"
-	"gopkg.in/yaml.v3"
 	sigsyaml "sigs.k8s.io/yaml"
 )
 
@@ -123,28 +118,6 @@ If an asset exists, it will be updated. If it doesn't exist, it will be created.
 	return cmd
 }
 
-// assetDocument represents a parsed YAML document with its kind
-type assetDocument struct {
-	kind     string
-	name     string // human-readable name extracted from the document
-	id       string // asset ID extracted from the document (location varies by kind)
-	raw      []byte
-	filePath string // relative path when loaded from a directory, empty for stdin/single-file
-	docIndex int    // 1-based index within the file
-	docCount int    // total number of documents in the file
-}
-
-// location returns a human-readable string describing where this document came from.
-func (d assetDocument) location() string {
-	if d.filePath == "" {
-		return fmt.Sprintf("document %d", d.docIndex)
-	}
-	if d.docCount == 1 {
-		return d.filePath
-	}
-	return fmt.Sprintf("%s: document %d", d.filePath, d.docIndex)
-}
-
 // applyAction indicates whether an asset was created or updated
 type applyAction string
 
@@ -164,13 +137,13 @@ type applyResult struct {
 }
 
 func runApply(ctx context.Context, flags *applyFlags) error {
-	var documents []assetDocument
+	var documents []asset.Document
 	var fromDirectory bool
 	var err error
 
 	if flags.File == "-" {
 		// Read from stdin
-		documents, err = readMultiDocumentYAML("-", os.Stdin)
+		documents, err = asset.ReadMultiDocumentYAML("-", os.Stdin)
 		if err != nil {
 			return validationError(err.Error())
 		}
@@ -181,12 +154,12 @@ func runApply(ctx context.Context, flags *applyFlags) error {
 		}
 		if info.IsDir() {
 			fromDirectory = true
-			documents, err = readDirectory(flags.File)
+			documents, err = asset.ReadDirectory(flags.File)
 			if err != nil {
 				return validationError(err.Error())
 			}
 		} else {
-			documents, err = readMultiDocumentYAML(flags.File, nil)
+			documents, err = asset.ReadMultiDocumentYAML(flags.File, nil)
 			if err != nil {
 				return validationError(err.Error())
 			}
@@ -232,7 +205,7 @@ func runApply(ctx context.Context, flags *applyFlags) error {
 		results, applyErr := applyDocument(ctx, apiClient, doc, dataset)
 		for _, r := range results {
 			displayKind := asset.KindDisplayName(r.kind)
-			label := formatNameAndId(r.name, r.id)
+			label := asset.FormatNameAndID(r.name, r.id)
 			applied = append(applied, fmt.Sprintf("%s %s", displayKind, label))
 
 			if r.action == actionUpdated && r.before != nil {
@@ -240,7 +213,7 @@ func runApply(ctx context.Context, flags *applyFlags) error {
 					return err
 				}
 			} else if fromDirectory {
-				fmt.Printf("%s: %s %s %s\n", doc.filePath, displayKind, label, r.action)
+				fmt.Printf("%s: %s %s %s\n", doc.FilePath, displayKind, label, r.action)
 			} else {
 				fmt.Printf("%s %s %s\n", displayKind, label, r.action)
 			}
@@ -252,7 +225,7 @@ func runApply(ctx context.Context, flags *applyFlags) error {
 					fmt.Printf("  - %s\n", a)
 				}
 			}
-			return fmt.Errorf("%s (%s): %w", doc.location(), doc.kind, applyErr)
+			return fmt.Errorf("%s (%s): %w", doc.Location(), doc.Kind, applyErr)
 		}
 	}
 
@@ -269,7 +242,7 @@ func runApply(ctx context.Context, flags *applyFlags) error {
 			if confirmErr != nil || !confirmed {
 				skipped := len(deletionPlan.plan.ByIdentifier) + len(deletionPlan.plan.AlertsByName)
 				fmt.Fprintf(os.Stderr, "--since's deletion phase skipped; the rest of the run already completed\n")
-				return fmt.Errorf("%s not confirmed for deletion (--since ref is not an ancestor of HEAD)", pluralize(skipped, "asset"))
+				return fmt.Errorf("%s not confirmed for deletion (--since ref is not an ancestor of HEAD)", asset.Pluralize(skipped, "asset"))
 			}
 		}
 		declined, err := applyDeletions(ctx, apiClient, dataset, deletionPlan, flags.Force)
@@ -277,7 +250,7 @@ func runApply(ctx context.Context, flags *applyFlags) error {
 			return err
 		}
 		if declined > 0 {
-			return fmt.Errorf("%s declined; the rest of the --since run completed", pluralize(declined, "deletion"))
+			return fmt.Errorf("%s declined; the rest of the --since run completed", asset.Pluralize(declined, "deletion"))
 		}
 	}
 
@@ -288,36 +261,36 @@ func runApply(ctx context.Context, flags *applyFlags) error {
 // never partially triggered by a problem detectable before the first API call. Non-fatal warnings
 // are collected separately — callers only print them when validation succeeds, since a warning
 // about a document that never gets applied would be noise next to a hard error.
-func validateDocuments(documents []assetDocument) (validationErrors, validationWarnings []string) {
+func validateDocuments(documents []asset.Document) (validationErrors, validationWarnings []string) {
 	for _, doc := range documents {
-		if doc.kind == "" {
-			validationErrors = append(validationErrors, fmt.Sprintf("%s: missing 'kind' field", doc.location()))
-		} else if !isValidKind(doc.kind) {
-			validationErrors = append(validationErrors, fmt.Sprintf("%s: unsupported kind %q (supported: Dashboard, PersesDashboard, CheckRule, PrometheusRule, SyntheticCheck, View, Dash0SpamFilter, Dash0NotificationChannel, Dash0Team)", doc.location(), doc.kind))
-		} else if normalizeKind(doc.kind) == "spamfilter" {
+		if doc.Kind == "" {
+			validationErrors = append(validationErrors, fmt.Sprintf("%s: missing 'kind' field", doc.Location()))
+		} else if !isValidKind(doc.Kind) {
+			validationErrors = append(validationErrors, fmt.Sprintf("%s: unsupported kind %q (supported: Dashboard, PersesDashboard, CheckRule, PrometheusRule, SyntheticCheck, View, Dash0SpamFilter, Dash0NotificationChannel, Dash0Team)", doc.Location(), doc.Kind))
+		} else if normalizeKind(doc.Kind) == "spamfilter" {
 			// Catch unknown spam filter apiVersions during validation rather
 			// than after the first PUT, so a partial apply of a multi-doc input
 			// is never triggered by a typo in apiVersion.
-			if _, err := asset.DetectSpamFilterAPIVersion(doc.raw); err != nil {
-				validationErrors = append(validationErrors, fmt.Sprintf("%s: %s", doc.location(), err.Error()))
+			if _, err := asset.DetectSpamFilterAPIVersion(doc.Raw); err != nil {
+				validationErrors = append(validationErrors, fmt.Sprintf("%s: %s", doc.Location(), err.Error()))
 			}
-		} else if normalizeKind(doc.kind) == "prometheusrule" {
+		} else if normalizeKind(doc.Kind) == "prometheusrule" {
 			// Catch CRDs that contain no usable rules at all up front, before
 			// any API call. ParseAsPrometheusAlertRules already rejects
 			// alert-only-empty CRDs, but a CRD with zero rules of either kind
 			// would otherwise slip through to applyDocument.
-			if err := validatePrometheusRule(doc.raw); err != nil {
-				validationErrors = append(validationErrors, fmt.Sprintf("%s: %s", doc.location(), err.Error()))
+			if err := validatePrometheusRule(doc.Raw); err != nil {
+				validationErrors = append(validationErrors, fmt.Sprintf("%s: %s", doc.Location(), err.Error()))
 			}
-		} else if normalizeKind(doc.kind) == "notificationchannel" {
+		} else if normalizeKind(doc.Kind) == "notificationchannel" {
 			// A document carrying spec.routing.assets gets a non-fatal warning — the API treats
 			// the field as API-managed and silently ignores it on write; the apply proceeds as
 			// usual. Parse errors are already caught during metadata extraction in
-			// readMultiDocumentYAML.
+			// asset.ReadMultiDocumentYAML.
 			var channel dash0api.NotificationChannelDefinition
-			if err := sigsyaml.Unmarshal(doc.raw, &channel); err == nil {
+			if err := sigsyaml.Unmarshal(doc.Raw, &channel); err == nil {
 				if warning := asset.RoutingAssetsWarning(&channel); warning != "" {
-					validationWarnings = append(validationWarnings, fmt.Sprintf("%s: %s", doc.location(), warning))
+					validationWarnings = append(validationWarnings, fmt.Sprintf("%s: %s", doc.Location(), warning))
 				}
 			}
 		}
@@ -328,317 +301,7 @@ func validateDocuments(documents []assetDocument) (validationErrors, validationW
 // validationError formats one or more validation issues into a consistent
 // "validation failed with N error/errors:" message.
 func validationError(issues ...string) error {
-	return fmt.Errorf("validation failed with %s:\n  %s", pluralize(len(issues), "error"), strings.Join(issues, "\n  "))
-}
-
-// pluralize returns "1 thing" or "N things" depending on count.
-// This logic is currently naive and assumes all plurals are formed by adding "s".
-// This happens to work for all our asset kinds, errors, documents and other situations
-// we use this logic.
-func pluralize(count int, singular string) string {
-	if count == 1 {
-		return fmt.Sprintf("%d %s", count, singular)
-	}
-	return fmt.Sprintf("%d %ss", count, singular)
-}
-
-// formatNameAndId returns a display string with name and optional ID.
-func formatNameAndId(name, id string) string {
-	if name != "" && id != "" {
-		return fmt.Sprintf("%q (%s)", name, id)
-	}
-	if name != "" {
-		return fmt.Sprintf("%q", name)
-	}
-	if id != "" {
-		return fmt.Sprintf("(%s)", id)
-	}
-	return ""
-}
-
-// parseDocumentHeader extracts the kind, human-readable name, and ID from
-// raw YAML bytes. It unmarshals into typed structs and delegates to the
-// Extract* functions in internal/asset/ so that apply and the per-asset
-// CRUD commands use the same extraction logic.
-func parseDocumentHeader(data []byte) (kind, name, id string, err error) {
-	kind, err = dash0yaml.DetectKind(data)
-	if err != nil {
-		return "", "", "", err
-	}
-
-	switch normalizeKind(kind) {
-	case "dashboard":
-		var dashboard dash0api.DashboardDefinition
-		if err := sigsyaml.Unmarshal(data, &dashboard); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = dash0api.GetDashboardName(&dashboard)
-		if name == "" {
-			name = dashboard.Metadata.Name
-		}
-		id = dash0api.GetDashboardID(&dashboard)
-
-	case "checkrule":
-		var rule dash0api.PrometheusAlertRule
-		if err := sigsyaml.Unmarshal(data, &rule); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = dash0api.GetCheckRuleName(&rule)
-		id = dash0api.GetCheckRuleID(&rule)
-
-	case "view":
-		var view dash0api.ViewDefinition
-		if err := sigsyaml.Unmarshal(data, &view); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = dash0api.GetViewName(&view)
-		id = dash0api.GetViewID(&view)
-
-	case "syntheticcheck":
-		var check dash0api.SyntheticCheckDefinition
-		if err := sigsyaml.Unmarshal(data, &check); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = dash0api.GetSyntheticCheckName(&check)
-		id = dash0api.GetSyntheticCheckID(&check)
-
-	case "prometheusrule":
-		// We only need metadata (name + ID) here; the Metadata struct has no
-		// time.Duration fields, so a partial unmarshal via sigsyaml is safe.
-		var partial struct {
-			Metadata dash0api.PrometheusRulesMetadata `json:"metadata"`
-		}
-		if err := sigsyaml.Unmarshal(data, &partial); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = partial.Metadata.Name
-		id = partial.Metadata.Labels[dash0api.LabelID]
-
-	case "persesdashboard":
-		var perses dash0api.PersesDashboard
-		if err := sigsyaml.Unmarshal(data, &perses); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = dash0api.GetPersesDashboardName(&perses)
-		id = dash0api.GetPersesDashboardID(&perses)
-
-	case "spamfilter":
-		// Metadata fields (name + dash0.com/id label) are identical across
-		// v1alpha1 and v1alpha2, so we use the v1alpha1 type for the header
-		// peek regardless of the document's apiVersion. The apiVersion-aware
-		// dispatch happens in applyDocument.
-		var filter dash0api.SpamFilter
-		if err := sigsyaml.Unmarshal(data, &filter); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = dash0api.GetSpamFilterName(&filter)
-		id = dash0api.GetSpamFilterID(&filter)
-
-	case "notificationchannel":
-		var channel dash0api.NotificationChannelDefinition
-		if err := sigsyaml.Unmarshal(data, &channel); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = dash0api.GetNotificationChannelName(&channel)
-		id = dash0api.GetNotificationChannelID(&channel)
-		if id == "" {
-			// Notification channels use origin as the upsert key. Surface it
-			// as the ID in dry-run/listing output so users can see which
-			// existing channel the document will replace.
-			id = dash0api.GetNotificationChannelOrigin(&channel)
-		}
-
-	case "team":
-		var team dash0api.TeamDefinitionV1Alpha1
-		if err := sigsyaml.Unmarshal(data, &team); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = dash0api.GetTeamDisplayName(&team)
-		if name == "" {
-			name = dash0api.GetTeamName(&team)
-		}
-		id = dash0api.GetTeamID(&team)
-		if id == "" {
-			// Teams use origin as the upsert key.
-			id = dash0api.GetTeamOrigin(&team)
-		}
-
-	default:
-		var raw map[string]any
-		if err := sigsyaml.Unmarshal(data, &raw); err != nil {
-			return "", "", "", fmt.Errorf("failed to decode document: %w", err)
-		}
-		name = yamlStringFromMap(raw, "metadata", "name")
-	}
-
-	return kind, name, id, nil
-}
-
-// yamlString traverses a nested map[string]any by the given keys and returns
-// the leaf value as a string, or "" if any key is missing or the value is not
-// a string.
-func yamlStringFromMap(m map[string]any, keys ...string) string {
-	for i, key := range keys {
-		val, ok := m[key]
-		if !ok {
-			return ""
-		}
-		if i == len(keys)-1 {
-			s, _ := val.(string)
-			return s
-		}
-		m, ok = val.(map[string]any)
-		if !ok {
-			return ""
-		}
-	}
-	return ""
-}
-
-// readMultiDocumentYAML splits a YAML stream into individual documents.
-// This is the only place that requires gopkg.in/yaml.v3 directly —
-// sigs.k8s.io/yaml doesn't provide a streaming decoder for multi-document YAML.
-func readMultiDocumentYAML(filePath string, stdin io.Reader) ([]assetDocument, error) {
-	var data []byte
-	var err error
-
-	if filePath == "-" {
-		data, err = io.ReadAll(stdin)
-		if err != nil {
-			return nil, fmt.Errorf("failed to read from stdin: %w", err)
-		}
-		if len(data) == 0 {
-			return nil, fmt.Errorf("no input provided on stdin")
-		}
-	} else {
-		data, err = os.ReadFile(filePath)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return parseMultiDocumentYAML(data)
-}
-
-// parseMultiDocumentYAML splits data on YAML document boundaries and parses
-// each into an assetDocument. Factored out of readMultiDocumentYAML so
-// callers that already have file content in memory (e.g. --since's
-// git-history name lookups, which read a blob via ReadFileAtRef rather than
-// a path on disk) can reuse the same parsing without a round trip through
-// the filesystem.
-func parseMultiDocumentYAML(data []byte) ([]assetDocument, error) {
-	var documents []assetDocument
-	decoder := yaml.NewDecoder(bytes.NewReader(data))
-
-	for {
-		var node yaml.Node
-		err := decoder.Decode(&node)
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse YAML:\n    %w", err)
-		}
-
-		// Skip empty documents
-		if node.Kind == 0 {
-			continue
-		}
-
-		// Re-encode the node to get the raw bytes for this document
-		var buf bytes.Buffer
-		encoder := yaml.NewEncoder(&buf)
-		encoder.SetIndent(2)
-		if err := encoder.Encode(&node); err != nil {
-			return nil, fmt.Errorf("failed to re-encode document: %w", err)
-		}
-		encoder.Close()
-
-		kind, name, id, err := parseDocumentHeader(buf.Bytes())
-		if err != nil {
-			return nil, err
-		}
-
-		documents = append(documents, assetDocument{
-			kind:     kind,
-			name:     name,
-			id:       id,
-			raw:      buf.Bytes(),
-			docIndex: len(documents) + 1,
-		})
-	}
-
-	// Handle single-document files without YAML document markers
-	if len(documents) == 0 && len(data) > 0 {
-		kind, name, id, _ := parseDocumentHeader(data)
-		if kind != "" {
-			documents = append(documents, assetDocument{
-				kind:     kind,
-				name:     name,
-				id:       id,
-				raw:      data,
-				docIndex: 1,
-			})
-		}
-	}
-
-	// Set docCount on all documents
-	for i := range documents {
-		documents[i].docCount = len(documents)
-	}
-
-	return documents, nil
-}
-
-// discoverFiles recursively finds all .yaml/.yml files under dirPath,
-// skipping hidden entries (names starting with '.').
-// Returns paths relative to dirPath, sorted lexicographically.
-func discoverFiles(dirPath string) ([]string, error) {
-	var paths []string
-	var hasNestedDirs bool
-	if err := filepath.WalkDir(dirPath, asset.FindNonHiddenYAMLFiles(dirPath, &paths, &hasNestedDirs)); err != nil {
-		return nil, fmt.Errorf("failed to scan directory: %w", err)
-	}
-
-	var files []string
-	for _, path := range paths {
-		rel, err := filepath.Rel(dirPath, path)
-		if err != nil {
-			return nil, err
-		}
-		files = append(files, rel)
-	}
-	if len(files) == 0 {
-		if hasNestedDirs {
-			return nil, fmt.Errorf("no .yaml or .yml files found in %s and nested directories", dirPath)
-		}
-		return nil, fmt.Errorf("no .yaml or .yml files found in %s", dirPath)
-	}
-	sort.Strings(files)
-	return files, nil
-}
-
-// readDirectory reads all YAML files from a directory recursively.
-func readDirectory(dirPath string) ([]assetDocument, error) {
-	files, err := discoverFiles(dirPath)
-	if err != nil {
-		return nil, err
-	}
-
-	var allDocs []assetDocument
-	for _, relPath := range files {
-		fullPath := filepath.Join(dirPath, relPath)
-		docs, err := readMultiDocumentYAML(fullPath, nil)
-		if err != nil {
-			return nil, fmt.Errorf("%s: %w", relPath, err)
-		}
-		for i := range docs {
-			docs[i].filePath = relPath
-		}
-		allDocs = append(allDocs, docs...)
-	}
-
-	return allDocs, nil
+	return fmt.Errorf("validation failed with %s:\n  %s", asset.Pluralize(len(issues), "error"), strings.Join(issues, "\n  "))
 }
 
 func isValidKind(kind string) bool {
@@ -646,17 +309,13 @@ func isValidKind(kind string) bool {
 }
 
 func normalizeKind(kind string) string {
-	// Normalize common variations
-	k := strings.ToLower(strings.ReplaceAll(kind, "-", ""))
-	k = strings.ReplaceAll(k, "_", "")
-	k = strings.TrimPrefix(k, "dash0")
-	return k
+	return asset.NormalizeKind(kind)
 }
 
-func applyDocument(ctx context.Context, apiClient dash0api.Client, doc assetDocument, dataset *string) ([]applyResult, error) {
-	switch normalizeKind(doc.kind) {
+func applyDocument(ctx context.Context, apiClient dash0api.Client, doc asset.Document, dataset *string) ([]applyResult, error) {
+	switch normalizeKind(doc.Kind) {
 	case "dashboard", "persesdashboard":
-		dashboard, err := dash0yaml.ParseAsDashboard(doc.raw)
+		dashboard, err := dash0yaml.ParseAsDashboard(doc.Raw)
 		if err != nil {
 			return nil, err
 		}
@@ -677,7 +336,7 @@ func applyDocument(ctx context.Context, apiClient dash0api.Client, doc assetDocu
 
 	case "syntheticcheck":
 		var check dash0api.SyntheticCheckDefinition
-		if err := sigsyaml.Unmarshal(doc.raw, &check); err != nil {
+		if err := sigsyaml.Unmarshal(doc.Raw, &check); err != nil {
 			return nil, fmt.Errorf("failed to parse SyntheticCheck: %w", err)
 		}
 		result, err := asset.ImportSyntheticCheck(ctx, apiClient, &check, dataset)
@@ -687,11 +346,11 @@ func applyDocument(ctx context.Context, apiClient dash0api.Client, doc assetDocu
 				AssetName: check.Metadata.Name,
 			})
 		}
-		return []applyResult{{kind: doc.kind, name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
+		return []applyResult{{kind: doc.Kind, name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
 
 	case "view":
 		var view dash0api.ViewDefinition
-		if err := sigsyaml.Unmarshal(doc.raw, &view); err != nil {
+		if err := sigsyaml.Unmarshal(doc.Raw, &view); err != nil {
 			return nil, fmt.Errorf("failed to parse View: %w", err)
 		}
 		result, err := asset.ImportView(ctx, apiClient, &view, dataset)
@@ -701,14 +360,14 @@ func applyDocument(ctx context.Context, apiClient dash0api.Client, doc assetDocu
 				AssetName: view.Metadata.Name,
 			})
 		}
-		return []applyResult{{kind: doc.kind, name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
+		return []applyResult{{kind: doc.Kind, name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
 
 	case "spamfilter":
 		return applySpamFilter(ctx, apiClient, doc, dataset)
 
 	case "notificationchannel":
 		var channel dash0api.NotificationChannelDefinition
-		if err := sigsyaml.Unmarshal(doc.raw, &channel); err != nil {
+		if err := sigsyaml.Unmarshal(doc.Raw, &channel); err != nil {
 			return nil, fmt.Errorf("failed to parse Dash0NotificationChannel: %w", err)
 		}
 		result, err := asset.ImportNotificationChannel(ctx, apiClient, &channel)
@@ -722,7 +381,7 @@ func applyDocument(ctx context.Context, apiClient dash0api.Client, doc assetDocu
 
 	case "team":
 		var team dash0api.TeamDefinitionV1Alpha1
-		if err := sigsyaml.Unmarshal(doc.raw, &team); err != nil {
+		if err := sigsyaml.Unmarshal(doc.Raw, &team); err != nil {
 			return nil, fmt.Errorf("failed to parse Dash0Team: %w", err)
 		}
 		result, err := asset.ImportTeam(ctx, apiClient, &team)
@@ -735,7 +394,7 @@ func applyDocument(ctx context.Context, apiClient dash0api.Client, doc assetDocu
 		return []applyResult{{kind: "Dash0Team", name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
 
 	default:
-		return nil, fmt.Errorf("unsupported kind: %s", doc.kind)
+		return nil, fmt.Errorf("unsupported kind: %s", doc.Kind)
 	}
 }
 
@@ -743,8 +402,8 @@ func applyDocument(ctx context.Context, apiClient dash0api.Client, doc assetDocu
 // PrometheusRule CRD documents go through applyPrometheusRule, which inspects
 // the rule kinds and dispatches to the check-rule or recording-rule endpoint
 // (or both, when the CRD is mixed).
-func applyCheckRule(ctx context.Context, apiClient dash0api.Client, doc assetDocument, dataset *string) ([]applyResult, error) {
-	rules, err := asset.ParseCheckRules(doc.raw)
+func applyCheckRule(ctx context.Context, apiClient dash0api.Client, doc asset.Document, dataset *string) ([]applyResult, error) {
+	rules, err := asset.ParseCheckRules(doc.Raw)
 	if err != nil {
 		return nil, err
 	}
@@ -774,8 +433,8 @@ func applyCheckRule(ctx context.Context, apiClient dash0api.Client, doc assetDoc
 // check-rule endpoint via the existing ParseAsPrometheusAlertRules path;
 // recording rules are dispatched to the recording-rule endpoint via a
 // recording-only copy of the CRD.
-func applyPrometheusRule(ctx context.Context, apiClient dash0api.Client, doc assetDocument, dataset *string) ([]applyResult, error) {
-	crd, err := parsePrometheusRuleCRD(doc.raw)
+func applyPrometheusRule(ctx context.Context, apiClient dash0api.Client, doc asset.Document, dataset *string) ([]applyResult, error) {
+	crd, err := parsePrometheusRuleCRD(doc.Raw)
 	if err != nil {
 		return nil, err
 	}
@@ -840,8 +499,8 @@ func validatePrometheusRule(data []byte) error {
 // applySpamFilter handles both v1alpha1 and v1alpha2 spam filter documents.
 // The apiVersion field on the document selects the schema; an unknown value
 // is rejected with the list of supported versions before any API call.
-func applySpamFilter(ctx context.Context, apiClient dash0api.Client, doc assetDocument, dataset *string) ([]applyResult, error) {
-	apiVersion, err := asset.DetectSpamFilterAPIVersion(doc.raw)
+func applySpamFilter(ctx context.Context, apiClient dash0api.Client, doc asset.Document, dataset *string) ([]applyResult, error) {
+	apiVersion, err := asset.DetectSpamFilterAPIVersion(doc.Raw)
 	if err != nil {
 		return nil, err
 	}
@@ -849,7 +508,7 @@ func applySpamFilter(ctx context.Context, apiClient dash0api.Client, doc assetDo
 	switch apiVersion {
 	case string(dash0api.SpamFilterApiVersionV1Alpha1V1alpha1):
 		var filter dash0api.SpamFilter
-		if err := sigsyaml.Unmarshal(doc.raw, &filter); err != nil {
+		if err := sigsyaml.Unmarshal(doc.Raw, &filter); err != nil {
 			return nil, fmt.Errorf("failed to parse v1alpha1 SpamFilter: %w", err)
 		}
 		result, importErr := asset.ImportSpamFilter(ctx, apiClient, &filter, dataset)
@@ -859,10 +518,10 @@ func applySpamFilter(ctx context.Context, apiClient dash0api.Client, doc assetDo
 				AssetName: dash0api.GetSpamFilterName(&filter),
 			})
 		}
-		return []applyResult{{kind: doc.kind, name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
+		return []applyResult{{kind: doc.Kind, name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
 	case string(dash0api.V1alpha2):
 		var filter dash0api.SpamFilterV1Alpha2
-		if err := sigsyaml.Unmarshal(doc.raw, &filter); err != nil {
+		if err := sigsyaml.Unmarshal(doc.Raw, &filter); err != nil {
 			return nil, fmt.Errorf("failed to parse v1alpha2 SpamFilter: %w", err)
 		}
 		result, importErr := asset.ImportSpamFilterV1Alpha2(ctx, apiClient, &filter, dataset)
@@ -872,7 +531,7 @@ func applySpamFilter(ctx context.Context, apiClient dash0api.Client, doc assetDo
 				AssetName: filter.Metadata.Name,
 			})
 		}
-		return []applyResult{{kind: doc.kind, name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
+		return []applyResult{{kind: doc.Kind, name: result.Name, id: result.ID, action: applyAction(result.Action), before: result.Before, after: result.After}}, nil
 	default:
 		// Unreachable: DetectSpamFilterAPIVersion only returns supported values or an error.
 		return nil, fmt.Errorf("unsupported spam filter apiVersion %q", apiVersion)
